Falling_Detection/i3d/requirements.txt
av==9.2.0
numpy==1.22.3
Pillow==9.1.1
scikit_learn==1.1.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.0
